gh build fix

+++ src/components/Profile/Profile.tsx
@@ -8,8 +8,8 @@ import RatedCards from '../RatedCards/RatedCards';
 
 function Profile() {
     const { user } = useSelector((state: any) => state.user);
-    const { data: favoriteMovies, refetch: refetchFavorites } = useGetListQuery({ listName: 'favorite/movies', accountId: user.id, sessionId: localStorage.getItem('session_id'), page: 1 });
-    const { data: watchlistMovies, refetch: refetchWatchlisted } = useGetListQuery({ listName: 'watchlist/movies', accountId: user.id, sessionId: localStorage.getItem('session_id'), page: 1 });
+    const { data: favoriteMovies, refetch: refetchFavorites } = useGetListQuery({ listName: 'favorite/movies', accountId: user.id, sessionId: typeof window !== 'undefined' ? localStorage.getItem('session_id') : null, page: 1 });
+    const { data: watchlistMovies, refetch: refetchWatchlisted } = useGetListQuery({ listName: 'watchlist/movies', accountId: user.id, sessionId: typeof window !== 'undefined' ? localStorage.getItem('session_id'): null, page: 1 });
 
     useEffect(() => {
         refetchFavorites();
@@ -17,7 +17,7 @@ function Profile() {
     }, []);
 
     const logout = () => {
-        localStorage.clear();
+        if (typeof window !== 'undefined') localStorage.clear();
         window.location.href = '/';
     };
